Extract integration tests.
